+++ src/context/DeleteContext.js
@@ -1,0 +1,43 @@
+import { createContext, useContext, useState } from "react";
+import axios from "axios";
+import { useNote } from "./NoteContext";
+
+const DeleteContext = createContext({});
+
+const DeleteProvider = ({ children }) => {
+  const { setNotes } = useNote();
+  const [deletedNotes, setDeletedNotes] = useState([]);
+
+  const moveToTrash = (note) => {
+    setDeletedNotes((prev) => [...prev, note]);
+  };
+  const deleteNoteApiCall = async (notesId, note) => {
+    const encodedToken = localStorage.getItem("token");
+    const config = {
+      headers: {
+        authorization: encodedToken,
+      },
+    };
+    try {
+      const response = await axios.delete(`/api/notes/${notesId}`, config);
+      setNotes(response.data.notes);
+      console.log(response.data.notes);
+    } catch (e) {
+      console.log(e);
+    }
+  };
+  return (
+    <DeleteContext.Provider
+      value={{
+        moveToTrash,
+        deletedNotes,
+        deleteNoteApiCall,
+        setDeletedNotes,
+      }}
+    >
+      {children}
+    </DeleteContext.Provider>
+  );
+};
+const useDeleteNote = () => useContext(DeleteContext);
+export { DeleteProvider, useDeleteNote };
