+++ src/context/index.js
@@ -1,0 +1,11 @@
+import { NoteProvider, useNote } from "./NoteContext";
+import { DeleteProvider, useDeleteNote } from "./DeleteContext";
+import { useAuth, AuthProvider } from "./AuthContext";
+export {
+  NoteProvider,
+  useNote,
+  DeleteProvider,
+  useDeleteNote,
+  useAuth,
+  AuthProvider,
+};
