+++ src/index.js
@@ -4,8 +4,9 @@ import "./index.css";
 import App from "./App";
 import { makeServer } from "./server";
 import { BrowserRouter as Router } from "react-router-dom";
-import { AuthProvider } from "./context/AuthContext";
-import { NoteProvider } from "./context/NoteContext";
+// import { AuthProvider } from "./context/AuthContext";
+// import { NoteProvider } from "./context/NoteContext";
+import { AuthProvider, NoteProvider, DeleteProvider } from "./context/index";
 
 // Call make Server
 makeServer();
@@ -15,7 +16,9 @@ ReactDOM.render(
     <Router>
       <AuthProvider>
         <NoteProvider>
-          <App />
+          <DeleteProvider>
+            <App />
+          </DeleteProvider>
         </NoteProvider>
       </AuthProvider>
     </Router>
